+++ src/main/java/warmup/two/WarmupTwoSecondSet.java
@@ -61,4 +61,24 @@ public class WarmupTwoSecondSet {
 
         return false;
     }
+
+    /**
+     * Given an array of ints, return true if the sequence of numbers 1, 2, 3 appears in the array somewhere.
+     *
+     * @param nums
+     * @return
+     */
+    public boolean array123(int[] nums) {
+        int numsLen = nums.length / 2;
+
+        if (nums.length < 3) return false;
+
+        for (int i = 0; i < numsLen+1; i++) {
+            if (nums[i] == 1 && nums[i+1] == 2 && nums[i+2] == 3) {
+                return true;
+            }
+        }
+
+        return false;
+    }
 }